requirements.txt
django==2.0.7

pymysql

pillow

redis

django-redis-sessions

djangorestframework

qiniu

django-redis
xlrd
xlwt